# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Suivi VtES API",
    "description": "API du suivi de pratique VtES (collection EN/FR, decks, parties, tournois). Pilote d'une PWA offline-first destinée à être portée sur barrins-project.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cartes": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "Recherche dans le catalogue",
        "description": "Cartes du catalogue VEKN, triées par nom. `q` cherche dans le nom anglais (sous-chaîne, sans tenir compte de la casse).",
        "operationId": "listCards",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 80
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CardCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "clan_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Clan Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardSummary"
                  },
                  "title": "Response Listcards"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/cartes/{card_id}": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "Fiche complète d'une carte",
        "operationId": "getCard",
        "parameters": [
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Card Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardRead"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bundles": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "Produits (précons, boîtes)",
        "operationId": "listBundles",
        "parameters": [
          {
            "name": "card_set_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Card Set Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 60
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BundleRead"
                  },
                  "title": "Response Listbundles"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bundles/{bundle_id}": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "Contenu d'un produit",
        "operationId": "getBundle",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bundle Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BundleContentRead"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/bundles/{bundle_id}/stock": {
      "post": {
        "tags": [
          "catalogue"
        ],
        "summary": "Verser un produit dans la collection",
        "description": "Ajoute le contenu du produit au stock, dans la langue indiquée. Les quantités s'additionnent à l'existant. Non idempotent : deux appels versent deux produits.",
        "operationId": "depositBundle",
        "parameters": [
          {
            "name": "bundle_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bundle Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleDeposit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardCopyRead"
                  },
                  "title": "Response Depositbundle"
                }
              }
            }
          },
          "404": {
            "description": "Ressource introuvable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec l'état des données (doublon, stock, règle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/langues": {
      "get": {
        "tags": [
          "catalogue"
        ],
        "summary": "Langues d'exemplaires",
        "operationId": "listLanguages",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LanguageRead"
                  },
                  "type": "array",
                  "title": "Response Listlanguages"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "catalogue"
        ],
        "summary": "Ajouter une langue",
        "operationId": "createLanguage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanguageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LanguageRead"
                }
              }
            }
          },
          "409": {
            "description": "Conflit avec l'état des données (doublon, stock, règle).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BundleContentRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "card_set_id": {
            "type": "integer",
            "title": "Card Set Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size",
            "description": "Nombre de cartes annoncé par le produit. Déclaratif : ne colle pas toujours exactement au contenu listé."
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "cards": {
            "items": {
              "$ref": "#/components/schemas/BundleCardRead"
            },
            "type": "array",
            "title": "Cards",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "card_set_id",
          "code"
        ],
        "title": "BundleContentRead",
        "description": "Contenu complet d'un produit.\n\nRéponse d'un futur `GET /bundles/{id}` (Lot 2). Deux usages la motivent :\nsavoir quel produit acheter pour compléter un deck, et verser ce contenu\ndans la collection — `cards` a exactement la forme qu'attend une création\nd'entrées de stock (carte + quantité), à laquelle il ne manque que la\nlangue de l'exemplaire acheté.\n\nLe service assemble ces lignes depuis les occurrences de type « precon »\ndu produit ; ce n'est pas une relation ORM directe, la carte se trouvant\nun cran plus loin, derrière l'impression."
      },
      "BundleDeposit": {
        "properties": {
          "language_code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Language Code"
          },
          "count": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Count",
            "description": "Nombre de produits identiques versés d'un coup.",
            "default": 1
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "language_code"
        ],
        "title": "BundleDeposit",
        "description": "Versement du contenu d'un produit dans la collection.\n\nLe produit fixe les cartes et leurs exemplaires ; il ne manque que la\nlangue de ce qui a été acheté, et le nombre de produits identiques."
      },
      "BundleRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "card_set_id": {
            "type": "integer",
            "title": "Card Set Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size",
            "description": "Nombre de cartes annoncé par le produit. Déclaratif : ne colle pas toujours exactement au contenu listé."
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          }
        },
        "type": "object",
        "required": [
          "id",
          "card_set_id",
          "code"
        ],
        "title": "BundleRead",
        "description": "Produit vendu dans une extension (deck préconstruit, boîte…).\n\nL'unité d'achat : ce qu'on vise pour compléter un deck, et ce qu'on verse\ndans la collection d'un seul geste. `code` peut être vide — 14 extensions\nn'ont qu'un précon, sans code propre."
      },
      "CardCategory": {
        "type": "string",
        "enum": [
          "crypt",
          "library"
        ],
        "title": "CardCategory",
        "description": "Les deux moitiés d'un deck VtES (CLAUDE.md §5)."
      },
...
      "CardCopyRead": {
        "properties": {
          "card_id": {
            "type": "integer",
            "title": "Card Id"
          },
          "language_code": {
            "type": "string",
            "title": "Language Code"
          },
          "quantity_owned": {
            "type": "integer",
            "title": "Quantity Owned",
            "description": "Nombre d'exemplaires réellement possédés."
          },
          "proxy_allowed": {
            "type": "boolean",
            "title": "Proxy Allowed",
            "description": "Autorise à jouer cette carte en proxy dans cette langue, sans la posséder. Une entrée avec 0 exemplaire possédé et le proxy autorisé est le cas normal d'une carte jouée en proxy."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "card": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CardSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "card_id",
          "language_code",
          "quantity_owned",
          "proxy_allowed"
        ],
        "title": "CardCopyRead",
        "description": "Exemplaires possédés d'une carte dans une langue."
      },
...
      "CardDisciplineRead": {
        "properties": {
          "discipline": {
            "$ref": "#/components/schemas/DisciplineRead"
          },
          "superior": {
            "type": "boolean",
            "title": "Superior",
            "description": "Niveau supérieur (code en majuscules chez krcg)."
          }
        },
        "type": "object",
        "required": [
          "discipline",
          "superior"
        ],
        "title": "CardDisciplineRead",
        "description": "Discipline requise par une carte, avec son niveau."
      },
...
      "CardRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "vekn_id": {
            "type": "integer",
            "title": "Vekn Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/CardCategory"
          },
          "clan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClanRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "group_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Code"
          },
          "advanced": {
            "type": "boolean",
            "title": "Advanced",
            "default": false
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "description": "Scan de la carte (version anglaise de référence)."
          },
          "sect": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SectRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "cost_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CostType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Nature du coût : une carte n'en a jamais qu'un seul."
          },
          "cost_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Value",
            "description": "Montant du coût. Chaîne et non entier : 25 cartes coûtent « X »."
          },
          "burn_option": {
            "type": "boolean",
            "title": "Burn Option",
            "default": false
          },
          "trifle": {
            "type": "boolean",
            "title": "Trifle",
            "default": false
          },
          "clan_requirement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clan Requirement",
            "description": "Clan(s) exigé(s) par la carte, séparés par une virgule."
          },
          "path_requirement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path Requirement"
          },
          "discipline_requirement": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DisciplineRequirement"
              },
              {
                "type": "null"
              }
            ],
            "description": "Manière de combiner les disciplines listées : une seule, au choix, ou toutes ensemble."
          },
          "card_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Text"
          },
          "flavor_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flavor Text"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "banned_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Banned On",
            "description": "Date de bannissement publiée par le VEKN, vide sinon."
          },
          "types": {
            "items": {
              "$ref": "#/components/schemas/CardTypeRead"
            },
            "type": "array",
            "title": "Types",
            "default": []
          },
          "discipline_links": {
            "items": {
              "$ref": "#/components/schemas/CardDisciplineRead"
            },
            "type": "array",
            "title": "Discipline Links",
            "default": []
          },
          "printings": {
            "items": {
              "$ref": "#/components/schemas/CardPrintingRead"
            },
            "type": "array",
            "title": "Printings",
            "default": []
          },
          "translations": {
            "items": {
              "$ref": "#/components/schemas/CardTranslationRead"
            },
            "type": "array",
            "title": "Translations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "vekn_id",
          "name",
          "category"
        ],
        "title": "CardRead",
        "description": "Fiche complète d'une carte du catalogue."
      },
...
      "CardSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "vekn_id": {
            "type": "integer",
            "title": "Vekn Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "$ref": "#/components/schemas/CardCategory"
          },
          "clan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClanRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "capacity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity"
          },
          "group_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Code"
          },
          "advanced": {
            "type": "boolean",
            "title": "Advanced",
            "default": false
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url",
            "description": "Scan de la carte (version anglaise de référence)."
          }
        },
        "type": "object",
        "required": [
          "id",
          "vekn_id",
          "name",
          "category"
        ],
        "title": "CardSummary",
        "description": "Vue courte d'une carte, pour les listes et l'autocomplétion."
      },
...
      "CardTypeRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "CardTypeRead",
        "description": "Type de carte."
      },
      "ClanRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "abbrev": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Abbrev"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ClanRead",
        "description": "Clan."
      },
      "CostType": {
        "type": "string",
        "enum": [
          "pool",
          "blood",
          "conviction"
        ],
        "title": "CostType",
        "description": "Nature du coût d'une carte de library.\n\nkrcg n'expose qu'un coût par carte (`{\"type\": ..., \"value\": ...}`) : une\ncarte ne coûte jamais à la fois du sang et du pool."
      },
...
      "DisciplineRequirement": {
        "type": "string",
        "enum": [
          "mono",
          "choice",
          "combo"
        ],
        "title": "DisciplineRequirement",
        "description": "Manière dont une carte de library combine ses disciplines requises.\n\n`MONO` : une seule discipline (ou aucune). `CHOICE` : l'une au choix\nparmi plusieurs. `COMBO` : toutes exigées ensemble. Remplace le champ\ntextuel brut qu'imposaient les CSV, où « / » et « & » portaient cette\nnuance."
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Explication lisible de l'échec."
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "description": "Corps d'une réponse d'erreur métier."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LanguageCreate": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 8,
            "minLength": 1,
            "title": "Code"
          },
          "label": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Label"
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "label"
        ],
        "title": "LanguageCreate",
        "description": "Ajout d'une langue."
      },
      "LanguageRead": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Code court de la langue (EN, FR, ES, XX…)."
          },
          "label": {
            "type": "string",
            "title": "Label",
            "description": "Libellé affichable."
          },
          "sort_order": {
            "type": "integer",
            "title": "Sort Order",
            "description": "Ordre d'affichage dans les sélecteurs."
          }
        },
        "type": "object",
        "required": [
          "code",
          "label",
          "sort_order"
        ],
        "title": "LanguageRead",
        "description": "Une langue d'exemplaire.\n\nListe ouverte servie par l'API plutôt que codée en dur dans le front : le\njeu exact des langues reste à confirmer (CLAUDE.md §11 point 2), et une\nlangue ajoutée en base doit apparaître dans l'UI sans redéploiement."
      },
...
      "SectRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "SectRead",
        "description": "Sect."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}